requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
python-docx
ollama
duckduckgo-search
jinja2
python-multipart
aiofiles
pypdf2
